补一些之前漏的

+++ Unity/Assets/Scripts/Base/Helper/FileHelper.cs
@@ -25,6 +25,8 @@ namespace ETModel
 		
 		public static void CleanDirectory(string dir)
 		{
+            if (!Directory.Exists(dir)) return;
+
 			foreach (string subdir in Directory.GetDirectories(dir))
 			{
 				Directory.Delete(subdir, true);
